Struct unify::Squash

pub struct Squash<P: ?Sized> {
    // some fields omitted
}

Evidence of a proposition where the witness has been forgotten

Methods

impl<P: ?Sized> Squash<P>

fn new<'a>(_: &'a P) -> Squash<P>

Construct a Squash given a reference to a witness